% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_lm.R
\name{hw3_lm}
\alias{hw3_lm}
\title{Fitting Linear Models}
\usage{
hw3_lm(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted}

\item{data}{an data frame containing the variables in the model.}
}
\value{
coefficients a named vector of coefficients

residuals the residuals

fitted.values the fitted mean values

call the matched call
}
\description{
hw3_lm is used to fit linear regression model and return a summary table including the coefficients, residuals, and R-squared of the model.
}
\examples{
hw3_lm(mpg ~ wt + hp + cyl, data = mtcars)

}
