% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_lm.R
\name{print.hw3_lm}
\alias{print.hw3_lm}
\title{print.hw3_lm}
\usage{
\method{print}{hw3_lm}(x, ...)
}
\arguments{
\item{x}{An object of class "hw3_lm".}

\item{...}{Additional arguments passed to the print function.}
}
\description{
Prints a summary of the hw3_lm object
}
