% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3_lm.R
\name{hw3_lm}
\alias{hw3_lm}
\title{hw3_lm}
\usage{
hw3_lm(formula, data)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted}

\item{data}{an data frame containing the variables in the model.}
}
\value{
coefficients

residuals

fitted.values
}
\description{
Fitting Linear Models
}
\examples{
hw3_lm(y~x1+x2,test_data)

}
